refactor: fix token code organization and circular dependency (fixes #237)
## Changes

1. **Move AstBuilder to building layer**
   - Moved `parsing/builder.rs` → `building/ast_builder.rs`
   - Fixes circular dependency: building no longer imports parsing
   - AstBuilder clearly belongs in building layer (converts IR → AST)

2. **Rename token utilities to normalization**
   - Renamed `building/token/` → `building/normalization/`
   - Clarifies purpose: input adaptation for AST building
   - Renamed `normalization.rs` → `normalize.rs`
   - Renamed `processing.rs` → `utilities.rs`

3. **Update module structure**
   - Updated `building.rs` to expose `ast_builder` and `normalization`
   - Updated `parsing.rs` to remove `builder` module
   - Created `normalization/mod.rs` with proper re-exports

4. **Fix import paths**
   - Updated `parsing/linebased/engine.rs` to import from `building::ast_builder`
   - Updated `building/pipeline.rs` to import from `building::ast_builder`
   - Updated `building/api.rs` to use `normalization` instead of `token`
   - Updated `building/extraction.rs` to use `normalization::utilities`

## Benefits

✅ Fixes circular dependency (building no longer imports parsing)
✅ Clear layer cake: lexing → parsing → building
✅ AstBuilder in correct layer
✅ Token normalization purpose is clearer
✅ All 499 tests pass

## Architecture

```
lexing/     → produces LineToken
parsing/    → builds LineContainer, produces ParseNode IR
building/   → converts ParseNode → AST (including AstBuilder)
```
