+++ lex-parser/src/lex/building/normalization/utilities.rs
