+++ lex-parser/src/lex/building/pipeline.rs
@@ -3,7 +3,7 @@
 //! Currently this is a thin wrapper around the `AstBuilder` that converts parse
 //! nodes produced by the analyzers into the final `Document` AST.
 
-use crate::lex::parsing::builder::AstBuilder;
+use crate::lex::building::ast_builder::AstBuilder;
 use crate::lex::parsing::ir::ParseNode;
 use crate::lex::parsing::Document;
