--- lex-parser/src/lex/building/token.rs
@@ -1,5 +1,0 @@
-//! Token processing utilities for AST building
-
-pub mod processing;
-
-pub(crate) mod normalization;
