+++ lex-parser/src/lex/parsing/linebased/engine.rs
@@ -12,9 +12,9 @@
 //! making it testable and maintainable independently.
 
 use super::{declarative_grammar, tree_builder};
+use crate::lex::building::ast_builder::AstBuilder;
 use crate::lex::lexing::tokens_core::Token;
 use crate::lex::lexing::tokens_linebased::LineContainer;
-use crate::lex::parsing::builder::AstBuilder;
 use crate::lex::parsing::ir::{NodeType, ParseNode};
 use crate::lex::parsing::Document;
 use std::ops::Range as ByteRange;
